hazelcast-cloud-ml-worker/requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
hazelcast-python-client==3.12.2
idna==2.9
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
pandas==1.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
urllib3==1.25.8